+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * 【core   】      优化IndexedComparator性能（pr#1240@Gitee）
 * 【http   】      改进ContentType.get忽略空格（pr#3664@Github）
 * 【http   】      CompressUtil.createExtractor支持tgz自动识别（pr#3674@Github）
+* 【poi    】      ExcelWriter.autoSizeColumn增加可选widthRatio参数，可配置中文字符宽度倍数（pr#3689@Github）
 
 ### 🐞Bug修复
 * 【core   】      修复因RFC3986理解有误导致的UrlPath处理冒号转义问题（issue#IAAE88@Gitee）